- updated news tableViewCell to use a CollectionView
